requirements.txt
aiogram>=3.13.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
colorama>=0.4.6
qrcode[pil]>=7.4.0
Pillow>=10.0.0
